+++ src/Loader/Hmr/hot-update.js
@@ -1,2 +1,2 @@
 // To enable live reload, just add an empty JS file in HTML.
-// Webpack automatically inserts the hot update code into this file.
+// Webpack automatically inserts the runtime code into this file.