+++ test/manual/watch-css-hmr-import-multiple-js/src/app.js
@@ -1,0 +1,7 @@
+import timer from './timer.js';
+
+import './timer.css';
+//import './style-a.css';
+//import './style-b.css';
+
+timer('#timer');