+++ test/manual/watch-css-hmr-import-single-js/src/app.js
@@ -1,0 +1,8 @@
+import timer from './timer.js';
+
+// import multiple styles in the single JS file => HMR works for all CSS files
+import './timer.css';
+import './style-a.css';
+import './style-b.css';
+
+timer('#timer');