+++ test/manual/watch-css-hmr-style-loader/src/script-a.js
@@ -1,0 +1,3 @@
+import './style-a.css';
+
+console.log('>> script-a');