+++ test/manual/watch-css-hmr-style-loader/src/script-b.js
@@ -1,0 +1,3 @@
+import './style-b.css';
+
+console.log('>> script-b');